move builder (#302)
